+++ src/main/rust/src/luajit_constants.rs
@@ -1,4 +1,4 @@
 
-pub const LUAJIT_MODE_OFF: int =		0x0000;	/* Turn feature off. */
-pub const LUAJIT_MODE_ON: int =		0x0100;	/* Turn feature on. */
-pub const LUAJIT_MODE_FLUSH: int =     0x0200;	/* Flush JIT-compiled code. */
+pub const LUAJIT_MODE_OFF: isize =		0x0000;	/* Turn feature off. */
+pub const LUAJIT_MODE_ON: isize =		0x0100;	/* Turn feature on. */
+pub const LUAJIT_MODE_FLUSH: isize =     0x0200;	/* Flush JIT-compiled code. */
